style(site): update instructions settings layout
